+++ src/main/java/com/example/lms/model/Teacher.java
@@ -4,7 +4,7 @@ import lombok.EqualsAndHashCode;
 
 @EqualsAndHashCode(callSuper = true)
 public class Teacher extends User{
-    Teacher(String name, int userId, String email, String password, String phoneNumber, String profilePic, String googleId) {
+    Teacher(String name, long userId, String email, String password, String phoneNumber, String profilePic, String googleId) {
         super(name, userId, email, password, phoneNumber, profilePic, googleId);
     }
 }
